session-2-mcp/requirements.txt
langchain>=0.3
langchain-openai>=0.2
langgraph>=0.2
openai>=1.0
rdkit>=2023.9
pubchempy>=1.0
python-dotenv>=1.0
requests>=2.31
flask>=3.0
langchain-mcp-adapters==0.1.0
mcp>=1.0
uvicorn>=0.30
gradio_client>=1.0
ipython>=8.12.0,<9.0.0
jupyterlab
ipywidgets